requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
boto3==1.34.0
requests==2.31.0
python-multipart==0.0.6